move hostname regex pattern function to frontend-utils
